feature: Split model to independent file
