+++ scripts/database/mysqlConnect.py
@@ -6,7 +6,7 @@ import os
 import inspect
 import configparser
 from util import configSectionMap
-from peewee import *
+from peewee import MySQLDatabase, Model
 
 # read database config file
 config = configparser.ConfigParser()
@@ -28,19 +28,6 @@ conn = MySQLDatabase(mDb, host=mHost, port=mPort, user=mUser, passwd=mPasswd)
 class BaseModel(Model):
     class Meta:
         database = conn
-
-
-class PornVideo(BaseModel):
-    id = AutoField()
-    source = CharField()
-    view_numbers = IntegerField()
-    video_id = CharField()
-    view_ratings = CharField()
-    video_title = CharField()
-    create_date = DateTimeField()
-
-    class Meta:
-        table_name = 'porn_videos'
 
 
 def connectMysql():
